✨ feat(sortby): refactored the module to have only one function (feat #190)
all the smaller functions have been merged into one function which can be switched through
specifying the type of factor to use. not yet tested but looks stable enough for a commit

BREAKING CHANGE: The against and colorspace parameters are now specified as options. the function
takes a string or the factor as the first argument instead of a collection

feat #190